Only remove 99-netbox.yml if /ansible/inventory is writable (#158)

+++ files/scripts/master/run.sh
@@ -23,7 +23,8 @@ if [[ -e /ansible/ara.env ]]; then
     source /ansible/ara.env
 fi
 
-if [[ ! -e /run/secrets/NETBOX_TOKEN ]]; then
+export ANSIBLE_INVENTORY=$ANSIBLE_DIRECTORY/inventory
+if [[ ! -e /run/secrets/NETBOX_TOKEN && -w $ANSIBLE_INVENTORY ]]; then
     rm -f /ansible/inventory/99-netbox.yml
 fi
 
@@ -31,8 +32,6 @@ export ANSIBLE_CONFIG=$ENVIRONMENTS_DIRECTORY/ansible.cfg
 if [[ -e $ENVIRONMENTS_DIRECTORY/$ENVIRONMENT/ansible.cfg ]]; then
     export ANSIBLE_CONFIG=$ENVIRONMENTS_DIRECTORY/$ENVIRONMENT/ansible.cfg
 fi
-
-export ANSIBLE_INVENTORY=$ANSIBLE_DIRECTORY/inventory
 
 if [[ -w $ANSIBLE_INVENTORY ]]; then
     rsync -a /ansible/group_vars/ /ansible/inventory/group_vars/
